add ScripArgOptions for the cluster client
