+++ java/client/src/main/java/glide/api/GlideClient.java
@@ -48,12 +48,9 @@ import glide.api.commands.ScriptingAndFunctionsCommands;
 import glide.api.commands.ServerManagementCommands;
 import glide.api.commands.TransactionsCommands;
 import glide.api.models.GlideString;
-import glide.api.models.Script;
 import glide.api.models.Transaction;
 import glide.api.models.commands.FlushMode;
 import glide.api.models.commands.InfoOptions.Section;
-import glide.api.models.commands.ScriptOptions;
-import glide.api.models.commands.ScriptOptionsGlideString;
 import glide.api.models.commands.SortOptions;
 import glide.api.models.commands.SortOptionsBinary;
 import glide.api.models.commands.function.FunctionRestorePolicy;
@@ -61,10 +58,8 @@ import glide.api.models.commands.scan.ScanOptions;
 import glide.api.models.configuration.GlideClientConfiguration;
 import glide.utils.ArgsBuilder;
 import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import lombok.NonNull;
 import org.apache.commons.lang3.ArrayUtils;
@@ -508,47 +503,6 @@ public class GlideClient extends BaseClient
             @NonNull GlideString cursor, @NonNull ScanOptions options) {
         GlideString[] arguments = new ArgsBuilder().add(cursor).add(options.toArgs()).toArray();
         return commandManager.submitNewCommand(Scan, arguments, this::handleArrayResponseBinary);
-    }
-
-    @Override
-    public CompletableFuture<Object> invokeScript(@NonNull Script script) {
-        if (script.getBinaryOutput()) {
-            return commandManager.submitScript(
-                    script, List.of(), List.of(), this::handleBinaryObjectOrNullResponse);
-        } else {
-            return commandManager.submitScript(
-                    script, List.of(), List.of(), this::handleObjectOrNullResponse);
-        }
-    }
-
-    @Override
-    public CompletableFuture<Object> invokeScript(
-            @NonNull Script script, @NonNull ScriptOptions options) {
-        if (script.getBinaryOutput()) {
-            return commandManager.submitScript(
-                    script,
-                    options.getKeys().stream().map(GlideString::gs).collect(Collectors.toList()),
-                    options.getArgs().stream().map(GlideString::gs).collect(Collectors.toList()),
-                    this::handleBinaryObjectOrNullResponse);
-        } else {
-            return commandManager.submitScript(
-                    script,
-                    options.getKeys().stream().map(GlideString::gs).collect(Collectors.toList()),
-                    options.getArgs().stream().map(GlideString::gs).collect(Collectors.toList()),
-                    this::handleObjectOrNullResponse);
-        }
-    }
-
-    @Override
-    public CompletableFuture<Object> invokeScript(
-            @NonNull Script script, @NonNull ScriptOptionsGlideString options) {
-        if (script.getBinaryOutput()) {
-            return commandManager.submitScript(
-                    script, options.getKeys(), options.getArgs(), this::handleBinaryObjectOrNullResponse);
-        } else {
-            return commandManager.submitScript(
-                    script, options.getKeys(), options.getArgs(), this::handleObjectOrNullResponse);
-        }
     }
 
     @Override
