+++ java/client/src/main/java/glide/api/commands/ScriptingAndFunctionsClusterCommands.java
@@ -5,12 +5,11 @@ import glide.api.models.ClusterValue;
 import glide.api.models.GlideString;
 import glide.api.models.Script;
 import glide.api.models.commands.FlushMode;
-import glide.api.models.commands.ScriptOptions;
-import glide.api.models.commands.ScriptOptionsGlideString;
+import glide.api.models.commands.ScriptArgOptions;
+import glide.api.models.commands.ScriptArgOptionsGlideString;
 import glide.api.models.commands.function.FunctionRestorePolicy;
 import glide.api.models.configuration.ReadFrom;
 import glide.api.models.configuration.RequestRoutingConfiguration.Route;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
@@ -1045,90 +1044,6 @@ public interface ScriptingAndFunctionsClusterCommands {
      * @see <a href="https://valkey.io/commands/script-load/">SCRIPT LOAD</a> and <a
      *     href="https://valkey.io/commands/evalsha/">EVALSHA</a> for details.
      * @param script The Lua script to execute.
-     * @return A value that depends on the script that was executed.
-     * @example
-     *     <pre>{@code
-     * try(Script luaScript = new Script("return 'Hello'", false)) {
-     *     String result = (String) client.invokeScript(luaScript).get();
-     *     assert result.equals("Hello");
-     * }
-     * }</pre>
-     */
-    CompletableFuture<Object> invokeScript(Script script);
-
-    /**
-     * Invokes a Lua script with its keys and arguments.<br>
-     * This method simplifies the process of invoking scripts on the server by using an object that
-     * represents a Lua script. The script loading, argument preparation, and execution will all be
-     * handled internally. If the script has not already been loaded, it will be loaded automatically
-     * using the <code>SCRIPT LOAD</code> command. After that, it will be invoked using the <code>
-     * EVALSHA</code> command.
-     *
-     * @apiNote When in cluster mode
-     *     <ul>
-     *       <li>all <code>keys</code> must map to the same hash slot.
-     *       <li>if no <code>keys</code> are given, command will be routed to a random primary node.
-     *     </ul>
-     *
-     * @see <a href="https://valkey.io/commands/script-load/">SCRIPT LOAD</a> and <a
-     *     href="https://valkey.io/commands/evalsha/">EVALSHA</a> for details.
-     * @param script The Lua script to execute.
-     * @param options The script option that contains keys and arguments for the script.
-     * @return A value that depends on the script that was executed.
-     * @example
-     *     <pre>{@code
-     * try(Script luaScript = new Script("return { KEYS[1], ARGV[1] }", false)) {
-     *     ScriptOptions scriptOptions = ScriptOptions.builder().key("foo").arg("bar").build();
-     *     Object[] result = (Object[]) client.invokeScript(luaScript, scriptOptions).get();
-     *     assert result[0].equals("foo");
-     *     assert result[1].equals("bar");
-     * }
-     * }</pre>
-     */
-    CompletableFuture<Object> invokeScript(Script script, ScriptOptions options);
-
-    /**
-     * Invokes a Lua script with its keys and arguments.<br>
-     * This method simplifies the process of invoking scripts on the server by using an object that
-     * represents a Lua script. The script loading, argument preparation, and execution will all be
-     * handled internally. If the script has not already been loaded, it will be loaded automatically
-     * using the <code>SCRIPT LOAD</code> command. After that, it will be invoked using the <code>
-     * EVALSHA</code> command.
-     *
-     * @apiNote When in cluster mode
-     *     <ul>
-     *       <li>all <code>keys</code> must map to the same hash slot.
-     *       <li>if no <code>keys</code> are given, command will be routed to a random primary node.
-     *     </ul>
-     *
-     * @see <a href="https://valkey.io/commands/script-load/">SCRIPT LOAD</a> and <a
-     *     href="https://valkey.io/commands/evalsha/">EVALSHA</a> for details.
-     * @param script The Lua script to execute.
-     * @param options The script option that contains keys and arguments for the script.
-     * @return A value that depends on the script that was executed.
-     * @example
-     *     <pre>{@code
-     * try(Script luaScript = new Script(gs("return { KEYS[1], ARGV[1] }", true))) {
-     *     ScriptOptionsGlideString scriptOptions = ScriptOptionsGlideString.builder().key(gs("foo")).arg(gs("bar")).build();
-     *     Object[] result = (Object[]) client.invokeScript(luaScript, scriptOptions).get();
-     *     assert result[0].equals(gs("foo"));
-     *     assert result[1].equals(gs("bar"));
-     * }
-     * }</pre>
-     */
-    CompletableFuture<Object> invokeScript(Script script, ScriptOptionsGlideString options);
-
-    /**
-     * Invokes a Lua script.<br>
-     * This method simplifies the process of invoking scripts on the server by using an object that
-     * represents a Lua script. The script loading and execution will all be handled internally. If
-     * the script has not already been loaded, it will be loaded automatically using the <code>
-     * SCRIPT LOAD</code> command. After that, it will be invoked using the <code>EVALSHA </code>
-     * command.
-     *
-     * @see <a href="https://valkey.io/commands/script-load/">SCRIPT LOAD</a> and <a
-     *     href="https://valkey.io/commands/evalsha/">EVALSHA</a> for details.
-     * @param script The Lua script to execute.
      * @param route Specifies the routing configuration for the command. The client will route the
      *     command to the nodes defined by <code>route</code>.
      * @return A value that depends on the script that was executed.
@@ -1150,12 +1065,6 @@ public interface ScriptingAndFunctionsClusterCommands {
      * using the <code>SCRIPT LOAD</code> command. After that, it will be invoked using the <code>
      * EVALSHA</code> command.
      *
-     * @apiNote When in cluster mode
-     *     <ul>
-     *       <li>all <code>keys</code> must map to the same hash slot.
-     *       <li>if no <code>keys</code> are given, command will be routed to a random primary node.
-     *     </ul>
-     *
      * @see <a href="https://valkey.io/commands/script-load/">SCRIPT LOAD</a> and <a
      *     href="https://valkey.io/commands/evalsha/">EVALSHA</a> for details.
      * @param script The Lua script to execute.
@@ -1171,7 +1080,7 @@ public interface ScriptingAndFunctionsClusterCommands {
      * }
      * }</pre>
      */
-    CompletableFuture<Object> invokeScript(Script script, List<String> args, Route route);
+    CompletableFuture<Object> invokeScript(Script script, ScriptArgOptions options, Route route);
 
     /**
      * Invokes a Lua script with its keys and arguments.<br>
@@ -1180,12 +1089,6 @@ public interface ScriptingAndFunctionsClusterCommands {
      * handled internally. If the script has not already been loaded, it will be loaded automatically
      * using the <code>SCRIPT LOAD</code> command. After that, it will be invoked using the <code>
      * EVALSHA</code> command.
-     *
-     * @apiNote When in cluster mode
-     *     <ul>
-     *       <li>all <code>keys</code> must map to the same hash slot.
-     *       <li>if no <code>keys</code> are given, command will be routed to a random primary node.
-     *     </ul>
      *
      * @see <a href="https://valkey.io/commands/script-load/">SCRIPT LOAD</a> and <a
      *     href="https://valkey.io/commands/evalsha/">EVALSHA</a> for details.
@@ -1202,7 +1105,8 @@ public interface ScriptingAndFunctionsClusterCommands {
      * }
      * }</pre>
      */
-    CompletableFuture<Object> invokeScriptBinary(Script script, List<GlideString> args, Route route);
+    CompletableFuture<Object> invokeScript(
+            Script script, ScriptArgOptionsGlideString options, Route route);
 
     /**
      * Checks existence of scripts in the script cache by their SHA1 digest.<br>
