+++ java/client/src/main/java/glide/api/models/commands/ScriptArgOptions.java
@@ -1,7 +1,7 @@
 /** Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api.models.commands;
 
-import glide.api.commands.ScriptingAndFunctionsBaseCommands;
+import glide.api.commands.ScriptingAndFunctionsClusterCommands;
 import glide.api.models.Script;
 import java.util.List;
 import lombok.Getter;
@@ -9,14 +9,14 @@ import lombok.Singular;
 import lombok.experimental.SuperBuilder;
 
 /**
- * Optional arguments for {@link ScriptingAndFunctionsBaseCommands#invokeScript(Script,
- * ScriptOptions)} command.
+ * Optional arguments for {@link ScriptingAndFunctionsClusterCommands#invokeScript(Script,
+ * ScriptArgOptions)} command.
  *
  * @see <a href="https://valkey.io/commands/evalsha/">valkey.io</a>
  */
 @SuperBuilder
-public final class ScriptOptions extends ScriptArgOptions {
+public class ScriptArgOptions {
 
-    /** The keys that are used in the script. */
-    @Singular @Getter private final List<String> keys;
+    /** The arguments for the script. */
+    @Singular @Getter private final List<String> args;
 }
