+++ java/client/src/main/java/glide/api/models/commands/ScriptArgOptionsGlideString.java
@@ -1,7 +1,7 @@
 /** Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api.models.commands;
 
-import glide.api.commands.ScriptingAndFunctionsBaseCommands;
+import glide.api.commands.ScriptingAndFunctionsClusterCommands;
 import glide.api.models.GlideString;
 import glide.api.models.Script;
 import java.util.List;
@@ -10,14 +10,14 @@ import lombok.Singular;
 import lombok.experimental.SuperBuilder;
 
 /**
- * Optional arguments for {@link ScriptingAndFunctionsBaseCommands#invokeScript(Script,
- * ScriptOptionsGlideString)} command.
+ * Optional arguments for {@link ScriptingAndFunctionsClusterCommands#invokeScript(Script,
+ * ScriptArgOptionsGlideString)} command.
  *
  * @see <a href="https://valkey.io/commands/evalsha/">valkey.io</a>
  */
 @SuperBuilder
-public class ScriptOptionsGlideString extends ScriptArgOptionsGlideString {
+public class ScriptArgOptionsGlideString {
 
-    /** The keys that are used in the script. */
-    @Singular @Getter private final List<GlideString> keys;
+    /** The arguments for the script. */
+    @Singular @Getter private final List<GlideString> args;
 }
